backend_yggdrasil/requirements2.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.6
daphne==4.2.1
Django==5.2.5
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djongo==1.3.7
dnspython==2.7.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
mongoengine==0.29.1
msgpack==1.1.1
mysql==0.0.3
mysql-connector-python==9.4.0
mysqlclient==2.2.7
numpy==2.2.6
pandas==2.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pymongo==3.11.4
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.4.0
requests==2.32.3
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.4.0
zope.interface==7.2
